backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
